00.09.101 dev
objects.tools now is an abstract class
