+++ .gitignore
@@ -41,6 +41,7 @@ pip-delete-this-directory.txt
 test.py
 assets/
 library/
+raw/
 
 
 # Unit test / coverage reports
